+++ internal/commands/database/session/cancel_test.go
@@ -1,4 +1,4 @@
-package databaseconnection
+package databasesession
 
 import (
 	"testing"
@@ -8,18 +8,19 @@ import (
 	smock "github.com/UpCloudLtd/upcloud-cli/v2/internal/mock"
 	"github.com/UpCloudLtd/upcloud-cli/v2/internal/mockexecute"
 
+	"github.com/UpCloudLtd/upcloud-go-api/v6/upcloud"
 	"github.com/UpCloudLtd/upcloud-go-api/v6/upcloud/request"
 	"github.com/stretchr/testify/assert"
 )
 
-func TestCreateCommand(t *testing.T) {
-	targetMethod := "CancelManagedDatabaseConnection"
+func TestCancelCommand(t *testing.T) {
+	targetMethod := "CancelManagedDatabaseSession"
 	uuid := "0fa980c4-0e4f-460b-9869-11b7bd62b833"
 	for _, test := range []struct {
 		name     string
 		args     []string
 		error    string
-		expected request.CancelManagedDatabaseConnection
+		expected request.CancelManagedDatabaseSession
 	}{
 		{
 			name:  "no process id",
@@ -29,7 +30,7 @@ func TestCreateCommand(t *testing.T) {
 		{
 			name: "soft cancel",
 			args: []string{"--pid", "123456"},
-			expected: request.CancelManagedDatabaseConnection{
+			expected: request.CancelManagedDatabaseSession{
 				UUID:      uuid,
 				Pid:       123456,
 				Terminate: false,
@@ -38,7 +39,7 @@ func TestCreateCommand(t *testing.T) {
 		{
 			name: "terminate",
 			args: []string{"--pid", "987654", "--terminate"},
-			expected: request.CancelManagedDatabaseConnection{
+			expected: request.CancelManagedDatabaseSession{
 				UUID:      uuid,
 				Pid:       987654,
 				Terminate: true,
@@ -49,6 +50,13 @@ func TestCreateCommand(t *testing.T) {
 			conf := config.New()
 			testCmd := CancelCommand()
 			mService := new(smock.Service)
+
+			mService.On("GetManagedDatabase", &request.GetManagedDatabaseRequest{UUID: uuid}).
+				Return(&upcloud.ManagedDatabase{
+					State: upcloud.ManagedDatabaseStateRunning,
+					Type:  upcloud.ManagedDatabaseServiceTypeMySQL,
+					UUID:  uuid,
+				}, nil)
 
 			mService.On(targetMethod, &test.expected).Return(nil)